seaperate db and pool , tidy up in different functional py
